Reference count for threads.
The reference count for a thread was added in the thread's internal
data structure so that the runtime could determine whether it is
safe to release the memory allocated for a thread.  Currently,
the thread's reference count is only increased when it is created
with its handle. The last argument for ABT_Thread_create indicates
the thread handle and if it is not NULL its reference count is
increased. That is, if users obtain thread's handle, it is their
responsibility to free the thread object through ABT_Thread_free().
However, if the execution stream associated with the thread is
freed via ABT_Stream_free(), its related threads are automatically
released even though there is no explicit ABT_Thread_free() call.