Fix a bug in ABT_thread_yield().
When there are requests for ULT, e.g., cancel, migration, etc.,
ABT_thread_yield() should switch context from the calling ULT to
the scheduler regardless of existence of work units in the
scheduler's pool, so that the scheduler can handle the requests.
Previously, this condition was not checked. This patch fixes the
issue.

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>