Add ABTD_xstream_context_set_affinity().
ABTD_xstream_context_set_affinity() is used to set the CPU affinity
for ES. Unless the user sets the environment variable, ABT_ENV_SET_AFFINITY,
to 0 or NO, each ES has the CPU affinity according to its rank.
For example, ES 0 is bound to CPU core 0, ES 1 to CPU core 1, and so on.

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>