Add --disable-affinity option to configure
When combining Argobots with other libraries such as MPI, the other
library may not want Argobots to set the affinity. For example, MPI will
start several processes on the same node, these processes do not want
its threads to be bound to the same core.

This commit adds the option '--disable-affinity' to configure, if used,
the affinity of ES will not be set even the platform support it.

Signed-off-by: Sangmin Seo <sseo@anl.gov>