Add ABT_self_get_type().
ABT_self_get_type() is for checking the type of work unit that calls
this routine.  If the caller of this routine is a ULT or tasklet,
it gets ABT_UNIT_TYPE_THREAD or ABT_UNIT_TYPE_TASK, respectively.
However, if an external thread calls this routine, e.g., pthread,
it obtains ABT_UNIT_TYPE_EXT.  This functionality will be useful
when Argobots and pthread are mixedly used in the code.
Thank Alex Brooks <brooks8@illinois.edu> for suggesting this
functionality.

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>