Start ES when it is created.
ESes except the primary ES were started when a work unit is pushed to one of
pools that are exclusive to the target ES.  However, this caused overhead on
ULT/tasklet creation because the creation routine has to check whether the
associated ES has been started or not.  This patch changes this behavior so
that a new ES is automatically started when it is created.

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>