sched: handle the case when a stacked scheduler has no work unit.
If the scheduler is a stacked one and there is no work unit in its
associated pools, we have to escape from the scheduling function.
The scheduler will be stopped if it is a tasklet type. However, if
the scheduler is a ULT type, we context switch to the parent scheduler.

No reviewer.