Push the primary (main) ULT to its associated pool.
The primary (main) ULT has been excluded from scheduling by not pushing it into
a pool. However, it made the Argobots code more complicated and added overhead
because the primary ULT had to be checked everywhere. The main idea of this
patch is basically to push the primary ULT to its associated pool and to treat
it like a normal ULT. This change also accompanied the reimplementation of
ABTI_xstream_set_main_sched.

No reviewer.