Wait until the BLOCK request is reset.
In ABTI_thread_set_ready, we should wait until the scheduler of the
blocked ULT resets the BLOCK request. Otherwise, the ULT can be pushed
to a pool here and be scheduled by another scheduler if it is pushed to
a shared pool.  Accordingly, we do not need to check it in
ABTI_xstream_schedule.

No reviewer.