Exit the test when ABT_ERR_FEATURE_NA is returned.
When an Argobots routine returns ABT_ERR_FEATURE_NA and it is passed to
ABT_test_error, we just finish the test, i.e., skip it, instead of
making it fail. In this case, the test will be marked as SKIP in the
result summary of 'make check'.

No reviewer.