Do not set producer/consumer info for a pool not associated to a scheduler.
When a pool is not associated with any scheduler, we do not set
producer/consumer information in the pool to allow the creator of the
pool to push work units before it is associated with a scheduler.
Consequently, this change makes it possible to create a PRIVATE pool and
to add work units before the pool is used by a scheduler.

Signed-off-by: Halim Amer <aamer@anl.gov>