thread: bypass the scheduler and block the caller on join.
This patch is, if possible, to block the caller ULT and to directly jump
to the target ULT in ABT_thread_join. When the target ULT finishes its
execution, it switches back to the ULT called ABT_thread_join. Since we
bypass the scheduler, it improves the performance by reducing the number
of context-switch.

For testing, --enable-ult-join-opt has been added to configure.

No reviewer.