self: add ABT_self_get_arg.
ABT_self_get_arg() returns the argument for the caller's work unit
function.  If the caller is a ULT, this routine returns the function
argument passed to ABT_thread_create() when the caller was created.
On the other hand, if the caller is a tasklet, this routine returns
the function argument passed to ABT_task_create().  If the caller is
the primary ULT or an external thread, NULL will be returned.

No reviewer.