xstream: set affinity when ES is created.
We used to set the CPU affinity of ES when it starts its execution.
In that way, it may overwrite the affinity that was set by the user,
since the user can set or change the affinity after he created the ES.
In this patch, we set the affinity right after we create the ES context.

No reviewer.