key: bug fix in ABT_key_free.
There was a bug in ABT_key_free, which happens when the main ULT uses
a key that has a destructor registered and the key is freed before
ABT_finalize() is called.  When the main ULT is freed in ABT_finalize(),
we tried to access the destructor and it caused a seg. fault.  This
patch fixes the bug by utilizing a reference counter and a freed flag.

No reviewer.