thread: wake up the joiner on cancellation.
When we cancel a ULT, if other ULT is blocked to join the canceled ULT,
we have to wake up the joiner ULT.  Since the cancel request is handled
by the scheduler, we should not context switch to the joiner ULT and
need to always wake it up.

No reviewer.