Change the first parameter name in ABT_error_get_str.
ABT_error_get_str used 'errno' as the first parameter name, but it
could cause a problem when 'errno' is declared as a macro in errno.h.
For example, on the MCS workstations errno is defined as

extern int *__errno_location (void) __THROW __attribute__ ((__const__));
#define errno (*__errno_location ())

This patch changes the parameter name from 'errno' to 'err' to avoid
this problem.  Thanks to John Jenkins <jenkins@mcs.anl.gov> for
reporting this bug.

No reviewer.