mem-pool: check the availability of MAP_ANONYMOUS and MAP_ANON.
MAP_ANONYMOUS or MAP_ANON is to map anonymous memory not associated
with any specific file and it is necessary to implement a memory pool.
However, both of them are not included in the standard mmap
specification.  Consequently, we have to check the availability of
MAP_ANONYMOUS and MAP_ANON at the configure time, and if neither of
them is available, we need to fall back to using malloc() instead of
using mmap().

Thanks to Rob Ross <rross@mcs.anl.gov> for reporting the issue where
MAP_ANONYMOUS is not defined in <sys/mman.h>.

No reviewer.