Generate a warning if ABT_test_get_cycles is supported.
When ABT_test_get_cycles was not implemented on some platforms, we used
to generate a compilation error, but it caused build failure of the
entire test suite.  Instead of generating a compilation error, this
patch lets the compiler generate a warning and adds an empty
implementation of ABT_test_get_cycles.

No reviewer.