Support the symbol visibility
This patch implements shared library best practice with regard to symbol
visibility.  Argobots is compiled to hide all function and variable
symbols, except those decorated with ABT_API_PUBLIC.  Compiler support
is tested for, and visibility disabled if not available.  A configure
flag, --enable-visibility, is added to control this option.

No reviewer.