Experimental affinity mapping for Intel KNC.
This is an experimental affinity mapping for Intel Xeon Phi (KNC).
It seems that the OS kernel on Xeon Phi numbers contiguous CPU IDs at
a single physical core and then moves to the next physical core.  This
numbering causes worse performance when we use a small number of
physical cores.  So, we set the ES affinity in a round-robin manner
from the view of physical cores, not logical cores.

No reviewer.