valgrind: fix the valgrind support
When Argobots is configured with --enable-valgrind, all ULT stacks allocated
from heap are registered to Valgrind so that Valgrind can correctly handle ULT
stacks.  These stacks are registered on creation and release when threads are
freed.  To release registered regions, Argobots needs to remember the map
between Valgrind's id and stack. It is currently implemented with a naive linked
list protected by a simple lock. The performance is poor, but it is acceptable.

Signed-off-by: Halim Amer <aamer@anl.gov>