abt.h: make opaque types unique
Make opaque ABT_xxx types unique by introducing opaque struct types, which is
helpful to find a potential bugs.

For example, the previous definition is as follows:
  typedef void *ABT_thread;
The new definition is as follows:
  struct ABT_thread_opaque; // no member variables.
  typedef struct ABT_thread_opaque *ABT_thread;

There are several benefits for development. For example, passing a variable of
ABT_thread* type to an argument of ABT_thread type is now warned by compiler.

Search for "Opaque Pointer" for details.