thread_attr: replace userstack with stacktype.
"userstack" was a member variable of ABTI_thread_attr representing whether the
ULT stack is allocated by user or not. This commit replaces it with "stacktype",
an integer-type variable which represents how the ULT stack is allocated,
improving expressiveness while it consumes the same amount of memory; userstack
was of ABT_bool type, which is "int".

Signed-off-by: Halim Amer <aamer@anl.gov>