thread_attr: replace macros with inline functions.
ABTI_THREAD_ATTR_INIT_MIG, ABTI_thread_attr_init, and ABTI_thread_attr_copy are
defined as macros, which is error-prone and loses readability. This commit
replace them with inline functions. It is also debug-friendly.

It does not affect performance with decent compilers; ABTI_thread_attr_init is
inlined so migratable argument will not appear in the optimized code.

Signed-off-by: Halim Amer <aamer@anl.gov>