log: move ABTI_LOG_SET_SCHED to ABTI_thread_context_xxx functions.
lp_ABTI_log->p_sched must be maintained so that it always points to the current
scheduler if the execution stream is running a scheduler and NULL when running
ULTs. Currently, callers need to call ABTI_LOG_SET_SCHED to update this value
before jumping to schedulers. However, this rule is very error-prone, and
sometimes this policy is inconsistent.

This commit moves most ABTI_LOG_SET_SCHED functions to
ABTI_thread_context_switch_xxx and ABTI_thread_finish_context_xxx functions
to improve maintainability.

The following is exceptional and thus ABTI_LOG_SET_SCHED is called explicitly.
- Main schedulers (invoked by the main thread, or Pthreads)
- Stackable tasklet schedulers

The following ABTI_LOG_SET_SCHED is removed since they are not necessary.
- ABTI_LOG_SET_SCHED in ABTI_xstream_run_unit: there is no reason to call
  ABTI_LOG_SET_SCHED here.
- ABTI_LOG_SET_SCHED in ABTI_xstream_launch_main_sched: ABTI_xstream_schedule
  will call ABTI_LOG_SET_SCHED very soon.

The following potential bug is fixed.
- ABTI_LOG_SET_SCHED in ABTI_xstream_schedule_thread: though p_thread->is_sched
  must be set if a stackable ULT scheduler is invoked, NULL is always set.