thread: move ABTI_local_set_thread and ABTI_local_set_task.
ABTI_lp_local->p_thread and p_task must be updated whenever the context is
changed to a ULT. The previous rule is as follows:
- thread-to-thread: the caller needs to update.
- thread-to-sched : no update.
- sched-to-thread : the caller needs to update.
- sched-to-sched  : if scheduler is invoked, the caller needs to updates it.
                    Otherwise, no update.

This rule is very complicated and difficult to follow. This commit moves most
ABTI_local_set_thread and _task functions into the context switching functions
so that developers do not take care of these values.

Specifically, the rule is changed as follows:
- xxx-to-thread: the caller needs to update.
- xxx-to-sched : no update.

Manual updates are required at the following places:
- Main thread ULT (not started by the context switch)
- tasklet invocation

These two changes are nontrivial.

1. TLS value management on sched-sched context switch

Management of these TLS values on sched-sched context switch (which happens only
when the stackable scheduler feature is enabled) is changed. These values were
only used to know the previous ULT after context switches in
ABTI_xstream_schedule_thread; the previous ULT may not be the same as one to
which the context has been switched. This is the case if the target ULT is a
non-scheduler ULT and bypasses a scheduler.

However, according to the stackable scheduler concept, schedulers are always
linearly stacked, so the previous stackable scheduler ULT cannot be different
from the ULT to which the context has been switched. This commit uses this
knowledge.

Note that, originally, p_thread and p_task are not properly managed in a
scheduler and thus any Argobots functions using ABTI_lp_local may not be called
in a custom scheduler. Actually, it does not work when the stackable scheduler
feature is turned off. After this change, it does not work even when the
stackable scheduler feature is enabled.

2. Removing ABTI_local_set_thread in ABT_thread_join

Now all the context switch functions always update ABTI_lp_local->p_thread if
the target ULT is of non-scheduler type. Such a special handling is not needed.
Note that this control comes from the scheduler-bypass path in
ABTDI_thread_terminate, which did not call ABTI_local_set_thread.