thread: fix ABTI_LOG_SET_SCHED argument in ABTDI_thread_terminate.
Whenever a ULT jumps to the scheduler, the caller needs to update the current
scheduler via ABTI_LOG_SET_SCHED. However, this update is conditional in
ABTDI_thread_terminate. This commit fixes a bug by removing the condition so
that it always sets the parent scheduler to lp_ABTI_log->p_sched.

Signed-off-by: Halim Amer <aamer@anl.gov>