fcontext: add x86_64-elf init_and_call_fcontext
This commit implements the x86-64-elf version of init_and_call_fcontext
and exposes this function. This function is corresponding to the function named
switch_ctx_RoC_invoke in the following paper:

Lessons Learned from Analyzing Dynamic Promotion for User-Level Threading
S. Iwasaki, A. Amer, K. Taura, and P. Balaji (SC '18)

This function is used in the future commits.

Signed-off-by: Halim Amer <aamer@anl.gov>