thread: implement the dynamic promotion technique
This commit implements the dynamic promotion technique.

The dynamic promotion technique performs better when a ULT invoked by a
scheduler does not switch contexts during its execution, finishes its work, and
goes back to the parent scheduler. Specifically, the dynamic promotion uses
a normal "return" procedure instead of user-level context switch when the ULT
resumes the parent scheduler.

To achieve this, the following changes are necessary:
- sched-to-thread
  If the target ULT does not have the context yet, run the target ULT on top of
  the scheduler with init_and_call_fcontext. If the target ULT does not yield
  during its execution, the scheduler can be resumed by "return".
  If the target ULT already has the context, use the normal context switch.
- thread-to-sched
  Promote the current thread if this thread is not. Note this function is not
  called when the dynamic promotion thread *returns* to its parent scheduler.
- thread-to-thread
  Apply promotion to both threads if they are not promoted yet, and jump to the
  target thread.
- sched-to-sched
  Nothing special is required, since the premise guarantees that the scheduler
  ULTs are normal (=fully fledged) threads.

ctx->fctx is used to check if whether a ULT is promoted or not.

See the paper for the details of the algorithm.

Caveat: only non-scheduler ULTs can be of dynamic promotion type.

Signed-off-by: Halim Amer <aamer@anl.gov>