global: ABT_initialized returns true only if Argobots is available.
ABT_initialized returns true even if Argobots has not been fully initialized;
ABT_initialized only checks gp_ABTI_global, which is set at the very beginning
of ABT_init.  This implementation is correct since the description says it
returns ABT_SUCCESS if ABT_init has been called, though it is useless.

This commit fixes this issue by introducing a new variable (g_ABTI_initialized),
which is set to TRUE if the Argobots execution environment is ready.

When ABT_finalize is executed concurrently, Argobots might be unavailable even
if ABT_initialized returns true.  This will be fixed in the future.

Signed-off-by: Halim Amer <aamer@anl.gov>