thread: add a memory barrier to a loop calling a user function
The while loop busy waits until u_is_in_pool returns ABT_TRUE, though there is
no guarantee that u_is_in_pool, which can be given by a user, is aware of memory
ordering.

This patch adds a memory barrier to fix this issue.  Note that this path is
rarely taken so the additional barrier does not affect performance.