local: declare lp_ABTI_local at the beginning of functions
This commit changes the way of loading lp_ABTI_local: all functions that use
lp_ABTI_local load lp_ABTI_local as p_local at the beginning of the functions.
This change is done in order to make a diff in the next commit small.

This change breaks Argobots because the same pointer of lp_ABTI_local is used
even if context switch happens in the function.  The next commit will fix it.