local: pass TLS to internal functions via an argument
Reading TLS is potentially heavier than reading a normal variable because it
needs to read a special register to obtain a thread ID (on some architectures).
This commit changes the code so that internal functions (ABTI_xxx and ABTD_xxx)
take TLS (i.e., p_local) as an argument; all the external functions (ABT_xxx)
and functions called via a function pointer (e.g., sched_run) read TLS only once
at the beginning of the function and pass it to the internal functions.

This change is necessary to fix the problem of TLS caching without hurting the
performance much.  In the future commit, Argobots will be changed to really read
the TLS (in an expensive way) only after user-level context switch.