thread: remove redundant variables.
When the thread migration feature is disabled, p_req_arg and lock in ABTI_thread
are never used.  However, even when ABT_CONFIG_DISABLE_MIGRATION is enabled,
they are initialized in ABT_thread_create() and add overheads.  This commit
removes these two redundant members when the migration feature is disabled.