sched: change to the algorithm of basic scheduler
This change checks for events after X ULTs rather than X passes
through all the pools. The core idea is that pop_count is a better
metric to check for an event periodically compared with previous
work_count (pop_count * num_pools) since this metric does not
depend on the number of pools. Also, this change removed run_cnt
variable and used unit variable as input for SCHED_SLEEP,
which makes the code a bit cleaner since another varaible is not
required.