error: clean up ABTI_CHECK macros
Previously, two implementations of ABTI_CHECK macros exist based on
ABT_CONFIG_DISABLE_ERROR_CHECK; one is check the value and go to fn_fail if
there is an error while the other is empty. This commit adds a new macro,
ABTI_IS_ERROR_CHECK_ENABLED, which becomes 1 if the error check feature is
enabled, and uses this compile-time constant value to distinguish two cases.

Such a compile-time constant is propagated at compile time and thus this value
is not evaluated at runtime, so there is no performance penalty.

There are two additional merits:
- Fewer #ifndef - #endif. Code is cleaner.
- Refer to all arguments and fn_fail even if the error check feature is disabled.
  This does not affect the runtime behavior but can suppress some compiler's
  "unused" warnings.