error: add type checks in ABTI_CHECK_xxx macros.
Since comparison with NULL does not check the type of the pointer, so the given
argument type might be different from what is intended. This commit casts a NULL
pointer to a specific type so that it causes a warning or an error if a given
pointer has a different type.