abtu: round up size in ABTU_malloc.
This commit rounds up size to the multiple of cache line size to avoid false
sharing between Argobots objects and user-allocated objects.

For example, consider sizeof(ABTI_obj) = 36, sizeof(user_obj) = 16, and cache
line size is 64. In this case, even if ABTI_obj is aligned to cache line size,
this cache line might be shared with user_obj if user_obj is not aligned
(i.e., malloc). When sizeof(ABTI_obj) = 64, it doesn't happen.

This does not significantly increase memory footprint in Argobots because all
Argobots objects are aligned if --enable_aligned_alloc is set anyway. This
approach needs to round up size, while the overhead of ADD and AND is tiny.