event: remove ABT_event_XXX() functions
Because they have not been maintained for long, this patch removes
ABT_event_XXX() functions from Argobots. They support two features: adaptive ES
creation and reduction for power capping and profiling, both of which are
disabled by default. Removal of them eliminates a lot of possible dependences on
external libraries and cleans up the logic. These features should be implemented
on top of Argobots (e.g., as examples of how Argobots can interact with external
components) in the future.