mutex: use ABTI_unit_id to identify mutex owner
A recursive mutex needs to keep track of its owner. In addition to tasklets and
ULTs, external threads (i.e., Pthreads) can be the owner when an external-thread
feature is enabled.  Since a mutex uses ABTI_unit * for p_owner, however,
a pthread handler is cast to ABTI_unit * and assigned to p_owner when an
external thread owns a mutex, which potentially causes an error if developers
assume p_owner points to a real ABTI_unit.

To address it, an owner should be managed by ABTI_unit_id, which is a pointer
that can distinguish tasklets, ULTs, and external threads.