mem: use ABTI_native_thread_id to identify page owner
A memory page needs to keep track of its owner, which can be either an execution
stream or an external thread. However, this owner is type of ABTI_xstream *, so
if an external thread is an owner, a pthread handler is cast to ABTI_xstream *
and stored to the owner variable.

To avoid such a dangerous cast, this patch introduces ABTI_native_thread_id,
which is a pointer that can distinguish execution streams and external threads.