xstream_context: make ABTD_xstream_context struct-type
Previously, ABTD_xstream_context was an alias of pthread_t, but it will have
member variables in the future commit.  This patch is the first step to achieve
it by making ABTD_xstream_context structure type that has a pthread_t member
variable.  Note that this patch does not change the logic.