log: remove ABTI_LOG_SET_SCHED
ABTI_LOG_SET_SCHED() updates a currently running scheduler, which is used to
get a scheduler ID in LOG_PRINT().  This updates are mandatory when a scheduler
is invoked, suspended, and stopped and thus are scattered across
several functions.  For better code maintenance, this patch removes it.

The future commit will revive this scheduler print functionality without an
ad-hoc flag management.