thread: remove ABTD_thread_finish_context_sched_to_thread/child_sched
ABTD_thread_finish_context_sched_to_thread() and _to_child_sched() finish the
currently running thread and execute "a child work unit" that is considered to
be scheduled by a scheduler.  This unnatural behavior is needed only when the
main scheduler of the primary execution stream jumps to the main thread.  The
two functions that have generic interfaces are removed and a new function,
ABTI_thread_finish_context_sched_to_main_thread(), handles the corner case.