thread: categorize context switching types by target relationship
The essence of having different ABTI_thread_context_switch_xxx_internal() is not
the type of source/target scheduling entity (i.e., ABTI_sched or ABTI_thread)
but the relationship between the source and the target (i.e., parent, child, or
sibling) especially in terms of dynamic promotion (and potentially management of
scheduler stack).  This commit clarifies it by creating functions based on the
source-target relationship.

This slightly changes the functionality of public context switching functions,
which requires a change in ABTI_xstream_update_main_sched().