sched: set p_xstream->p_thread and p_task in a scheduler context
Previously, p_local_xstream->p_thread and p_task are managed only when an
underlying work unit is not a scheduler.  This is certainly efficient in terms
of the number of pointer updates, but it forces non-consistent pointer updates
based on an implicit special rule (i.e., whether a caller or a callee updates a
flag, etc).  It also forbids most ABT_self() functions in a scheduler context,
which is unreasonable.  For better code maintenance and better functionality in
a scheduler context, they should be managed even in a scheduler context.