sched: use a ULT for a tasklet-type stackable scheduler
The current implementation allows a tasklet-type stackable scheduler, but it is
very restrictive though it can reduce tiny overheads since schedulers are not
invoked frequently.  A tasklet-type stackable scheduler is modified to use a ULT
internally, while it keeps the external interface to specify a tasklet type.

This does not break the existing API and its semantics since previously there
was no way to get an underlying work unit implementation in a scheduler context.