log: replace ABTI_log_debug with ABTI_log_event
This patch removes ABTI_log_event() since no function refers to it.  The printed
information of ABTI_log_event() is more useful (e.g., information of the
underlying UTL/tasklets and ESs), however.  The implementation of
ABTI_log_debug() is therefore replaced by that of ABTI_log_event().  __FILE__
and __LINE__ will not be printed since they are not useful to debug this
relatively small threading library.