mem_pool: implement ABTI_mem_pool, a generic memory pool
ABTI_mem_pool is a generic memory pool implementation.  The basic algorithm is
similar to the current memory pool implementation; first it accesses a local
memory pool and then a global memory pool if it is empty or full.  The
advantages of the new algorithm are as follows:
- Generic: the implementation takes a segment size as a runtime argument.
- Per-bucket operation: not a single memory segment but multiple segments in a
  "bucket" are used when accessing a global memory pool, which reduces a number
  of global pool accesses.  The number of local buckets and the bucket size are
  constant, so the local pool do not keep too many segments.
- Lock-free: entire push-pop operations are lock-free (on most CPUs).  No ABA
  problem.

The current memory pool implementations should be replaced by this.