thread: remove ABTI_thread_req_arg
In the current implementation, ABTI_thread_req_arg is used only for thread
migration; req_arg holds a pool pointer to which a thread is migrated.  This
does not need to be a linked list, however, since only one migration request can
be processed per ULT; the old migration request is overwritten by the new one
(see ABTI_thread_add_req_arg() in the current implementation).  ABTI_thread
just needs to have a pool pointer to possess this pool information.