thread: remove a lock for migration
ABTI_thread has a lock to protect p_migration_pool, but because of its
overwriting nature, p_migration_pool just needs to hold a valid pointer when
the migration request is processed by the thread.  For this purpose, we just
need to update p_migration_pool before setting the migration request.  The
costly lock is unnecessary.