thread: clarify the semantics of thread migration
Thread migration requires asynchronous communication, but the descriptions of
the migration functions do not explain what users must guarantee and what the
runtime can take care of.  Specifically, it was not clear when the destination
component can be released.  This commit clarifies it by adding an explanation.
To follow the new specification, ABTI_thread_migrate_to_xstream() is changed to
migrate a thread only to a pool of the main scheduler (i.e., not a stacked
scheduler).

Note that this change will increase users' burden and allow runtime
optimizations.