thread: remove ABTI_xstream.sched_lock and ABTI_sched.state
Both ABTI_xstream.sched_lock and ABTI_sched.state are used to maintain schedulers
active during the migration process.  This synchronization mechanism is
complicated and costly; this should be guaranteed by users according to the
"clarified" descriptions of the migration functions.  This patch removes both.

This pacth could incur an issue if one uses the migration feature with stackable
schedulers or in a program that dynamically forks and joins execution streams.