examples: implement a profiler as an example of the tool feature
This patch implements a quite practical profiler as an example of the new tool
feature.  This profiler, called ABTX profiler (not ABT profiler because it is
not embedded in Argobots), is a header library that can print out all the
basic performance numbers including non-scheduling time and thread granularity.
Please see examples/profiling/abtx_prof.h for details.

Note that this profiler is not for a pure demonstration purpose; this profiler
employs several tricky techniques to reduce overheads.