examples/profiling: improve fancy printing format
ABTXI_prof_digit() considered a very small value to be a zero-digit number, which
can hide other small values that have a few digits after the decimal point.  For
example, an array of 0, 0.0001, and 0.0002 are printed as 0, 0, 0 (though
0, 0.0001, and 0.0002 are expected) since the fancy format tries to align values
with the largest digit (in this case, zero).  This patch fixes it by considering
a very small value as -99 digits number.