examples/profiling: print meaningful minimum values
With FANCY + DETAILED, the profiler shows several performance numbers of work
units (i.e., ULTs and tasklets).  However, if either of them is never used,
the minimum numbers always get zero because the minimum values of one of
them, for example, the minimum granularity of tasklets, is zero because no
tasklet is executed.  Such a number is meaningless.  This patch fixes this bug.