log: add a new option to discard created log messages
--enable-debug=log and --enable-debug=all generate gigantic log file even if a
running program is a small test.  This significantly slows down testing although
the actual log messages are unimportant.  This patch adds new hidden
configuration options, log-discard and all-discard: they create log messages
but do not print.  These options are useful to see if log message generation
causes a SEGV or not.

Note that this option is useless for Argobots users.