sched: forcibly free the primary scheduler on ABT_finalize()
It is extremely cumbersome to free a non-automatic scheduler that is running on
the primary execution stream; to do so, programmers first register a dummy
scheduler to the primary execution stream and then release the scheduler.  This
patch alleviates this burden by making ABT_finalize() always free the scheduler
of the primary execution stream and its associated pools.