affinity: refactor abtd_affinity.c
This patch introduces ABTD_affinity_cpuset to clean up the ABTD_affinity
interface.  This change removes the experimental ABT_AFFINITY_TYPE interface
(because this interface is too specific and thus should be replaced by a generic
affinity interface introduced in the future commit) and the special code paths
for FreeBSD (because the code paths for Linux/OSX also work on FreeBSD).