thread: remove a revival path in ABTI_thread_create_sched
When a stackable scheduler is enabled, ABTI_thread_create_sched() called in
ABT_pool_add_sched() tries to revive a thread if sched is reused.  However,
a thread may not be revived because ABTI_sched_free() frees it.  This patch
removes the revival path in ABTI_thread_create_sched() and lets it always create
a thread.