configure: always enable a stackable scheduler
Since there is little performance disadvantage to support a stackable scheduler,
this patch removes an option to disable a stackable scheduler (which was enabled
by default).  Note that this change seemingly increases a branch in
ABTI_thread_create_internal(), but this branch should be removed if
ABTI_thread_create_internal() is inlined properly.