thread: make variables for the migration feature optional
f_migration_cb, p_migration_cb_arg, and p_migration_pool are always allocated
with ABTI_thread by default.  This is wasteful since the migration feature is
rarely used.  This patch puts them in ABTI_thread->p_keytable so that we can
avoid allocation and initializion when unnecessary.

This degrades the performance of migration.