abti: introduce ABTI_IS_EXT_THREAD_ENABLED
ABTI_IS_EXT_THREAD_ENABLED is set to 1 if the external thread feature is
enabled; otherwise it is set to 0.  This macro can remove ugly ifdef and simpliy
the code structure.  Note that ABTI_IS_EXT_THREAD_ENABLED is evaluated at
compile time so it does not increase the overheads.