error: clarify execution contexts
This patch classifies execution contexts of all the Argobots functions into
the following three types:
 1. no specific context (assuming that Argobots has been already initialized),
 2. on an execution stream (i.e., not on an external thread), and
 3. on a yieldable thread (ULT).
This patch slightly changes the semantics of existing functions. For example,
there were functions that can be called by either an external thread or a ULT
but not a tasklet.  Such functions are updated so that they can be run on a
tasklet we well (though a tasklet will block the underlying execution stream).