xstream: introduce a root thread
A root thread is a thread that schedules a main scheduler.  Since a root thread
works as a parent of a main scheduler thread, a main scheduler thread can now
call synchronization operations that might call yield (e.g., ABT_mutex_lock()).
This change makes a user-visible non-tasklet ABT_thread (=a yieldable thread)
really yieldable.  Note that the root thread does not have a parent so it cannot
yield, but a root thread is not visible to users so it will not affect the API.